requirements.txt
spotipy==2.23.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1

